% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesma.R
\name{.apply_cluster}
\alias{.apply_cluster}
\title{Parallel job process using Snowball (Internal method with specific cluster
configuration method).}
\usage{
.apply_cluster(cl, x, fun, ..., .quiet = FALSE)
}
\arguments{
\item{cl}{a \code{Snowball cluster object}.}

\item{x}{a \code{list} of objects to be processed.}

\item{fun}{a \code{function} to be executed in a parallel way based on the data
defined.}

\item{...}{extra options for the job function handler.}
}
\value{
a \code{list} with the description of the executed job.
}
\description{
This function applies a data into a parallel function.
}
