% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesma.R
\name{do_parallel_jobs}
\alias{do_parallel_jobs}
\title{Parallel job process using Snowball (Internal method with specific cluster
configuration method).}
\usage{
do_parallel_jobs(file_out, jobs, multicores = 4)
}
\arguments{
\item{cl}{a \code{Snowball cluster object}.}

\item{x}{a \code{list} of objects to be processed.}

\item{fun}{a \code{function} to be executed in a parallel way based on the data
defined.}

\item{no_data_out}{a \code{numeric} value to used as the default no data value
in the generated raster.}

\item{gdal_options}{a \code{character} with extra options for the GDAL.}

\item{data_factor}{a \code{numeric} value defining the multiplication factor
for the image pixel values.}

\item{endmember}{a \code{data.frame} with the endmembers.}
}
\value{
a \code{list} with the description of the executed job.
}
\description{
This function generates mesma job objects.
}
