% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endmember.R
\name{as.endmember.data.frame}
\alias{as.endmember.data.frame}
\title{Transform a endmember list in a Data Frame}
\usage{
as.endmember.data.frame(endmember_definition)
}
\arguments{
\item{endmember_definition}{a \code{list} with the endmembers}
}
\value{
a \code{data.frame} with the endmembers defined.
}
\description{
This function transforms an endmember named list into a
data.frame
}
