% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesma.R
\name{make_jobs}
\alias{make_jobs}
\title{Generate job objects.}
\usage{
make_jobs(
  f,
  dtype_out = "Int16",
  no_data_out = -9999,
  gdal_options = "COMPRESS=LZW",
  data_factor = 10000,
  endmember = NULL
)
}
\arguments{
\item{f}{a \code{character} with the path to the image to be processed.}

\item{dtype_out}{a \code{character} describing the output data type.}

\item{no_data_out}{a \code{numeric} value to used as the default no data value
in the generated raster.}

\item{gdal_options}{a \code{character} with extra options for the GDAL.}

\item{data_factor}{a \code{numeric} value defining the multiplication factor
for the image pixel values.}

\item{endmember}{a \code{data.frame} with the endmembers.}
}
\value{
a \code{list} with the description of the executed job.
}
\description{
This function generates mesma job objects.
}
