% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesma.R
\name{do_mixture_model}
\alias{do_mixture_model}
\title{Parallel worker unit function to generate fraction image.}
\usage{
do_mixture_model(job, em)
}
\arguments{
\item{job}{a \code{list} with the definitions of the worker job to be executed
in the function.}

\item{em}{a \code{data.frame} with the definitions of the endmembers.}
}
\value{
a \code{list} with the description of the executed job.
}
\description{
This function calculates a fraction image for a piece of a
Raster Brick.
}
