% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endmember.R
\name{as.endmember.data.frame}
\alias{as.endmember.data.frame}
\title{Transform a endmember list in a Data Frame}
\usage{
as.endmember.data.frame(endmember_definition)
}
\arguments{
\item{...}{...}
}
\value{
todo
}
\description{
description
}
