% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesma.R
\name{mesma}
\alias{mesma}
\title{Crop Raster files by shapefile extents}
\usage{
mesma(raster_in, raster_out, endmember, ..., multicores = 4)
}
\arguments{
\item{...}{...}
}
\value{
todo
}
\description{
description
}
